api/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.6
django-allauth==0.52.0
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
filelock==3.9.0
gunicorn==19.9.0
idna==3.4
oauthlib==3.2.2
pipenv==2023.2.4
platformdirs==3.0.0
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
PyJWT==1.7.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
virtualenv==20.19.0
virtualenv-clone==0.5.7
